02_implementation/requirements.txt
# Core Framework
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# External APIs
tavily-python>=0.3.0
requests>=2.31.0

# State Management & Data
pydantic>=2.0.0
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0  # Retry logic
httpx>=0.25.0  # Async HTTP
aiohttp>=3.9.0  # Async requests

# Logging & Monitoring
structlog>=23.2.0
rich>=13.7.0  # Beautiful console output

# Web Interface
streamlit>=1.31.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
mypy>=1.7.0
